requirements.txt
numpy==1.24.3 
pandas==2.0.3 
scikit-learn==1.3.0 
matplotlib==3.7.1 
seaborn==0.12.2 
torch==2.0.1 
tensorflow==2.13.0 
scipy==1.11.1 
pyyaml==6.0 
jupyter==1.0.0 
liac-arff==2.5.0 
xgboost==1.7.6
lightgbm==4.0.0
catboost==1.2.2
